x86: Use __always_inline on ifunc auxiliary functions
clang does correctly propagate the no_stack_protector attribute for
inline function used in ifunc selectors.  For instance, the exp function
is expanded as:

  static inline void *
  ieee754_exp_ifunc_selector (void)
  {
    [...]
  }

  extern __typeof (__redirect_ieee754_exp) __ieee754_exp
    __attribute__ ((ifunc ("__ieee754_exp" "_ifunc")));

  static __attribute__((no_stack_protector)) __typeof (__redirect_ieee754_exp) *i
  __ieee754_exp_ifunc (void)
  {
    __typeof (__redirect_ieee754_exp) *res = ieee754_exp_ifunc_selector ();
    return res;
  };

The ifunc selection ended being compiled with -fstack-protector, and
with debug function set to the ifunc selector
(ieee754_exp_ifunc_selector) auxiliary function instead of the
selector itself.

The ifunc selector are called early, before TCB is setup, which
causes SIGSEGV for statically linked binaries.

To fix it use __always_inline instead of inline.  It should not change
code generation on gcc and it fixes the code generation issue on
clang.  It also fixes function debug information.